Complete restructuring.
